Updated to add Perlin functionality

+++ packages/contracts/src/systems/ship/MoveSystem.sol
@@ -8,14 +8,14 @@ import { CHUNK_SIZE } from "../../constants.sol";
 
 contract MoveSystem is System {
 
-    function move(uint256 newPositionX, uint256 newPositionY) public {
-        uint256 currentPositionX = Position.getPos_x(keccak256(abi.encodePacked(msg.sender)));
-        uint256 currentPositionY = Position.getPos_y(keccak256(abi.encodePacked(msg.sender)));
+    function move(int256 newPositionX, int256 newPositionY) public {
+        int256 currentPositionX = Position.getPos_x(keccak256(abi.encodePacked(msg.sender)));
+        int256 currentPositionY = Position.getPos_y(keccak256(abi.encodePacked(msg.sender)));
 
-        uint256 currentChunkX = currentPositionX / CHUNK_SIZE;
-        uint256 currentChunkY = currentPositionY / CHUNK_SIZE;
-        uint256 newChunkX = newPositionX / CHUNK_SIZE;
-        uint256 newChunkY = newPositionY / CHUNK_SIZE;
+        int256 currentChunkX = currentPositionX / CHUNK_SIZE;
+        int256 currentChunkY = currentPositionY / CHUNK_SIZE;
+        int256 newChunkX = newPositionX / CHUNK_SIZE;
+        int256 newChunkY = newPositionY / CHUNK_SIZE;
 
         // Make sure that this update is to a neighboring chunk or same chunk and not a move across the map.
         // In practice forces players to update their position consistenly over time like a update function.
@@ -41,18 +41,18 @@ contract MoveSystem is System {
         // TODO: Updates for when leaving the port
     }
 
-    function sqrt(uint256 x) private pure returns (uint256) {
+    function sqrt(int256 x) private pure returns (uint256) {
         if (x == 0) {
             return 0;
         }
 
-        uint256 z = (x + 1) / 2;
-        uint256 y = x;
+        int256 z = (x + 1) / 2;
+        int256 y = x;
         while (z < y) {
             y = z;
             z = (x / z + z) / 2;
         }
-        return y;
+        return uint256(y);
     }
 
 }
